Fixed logout on ticket page bug

+++ react-app/src/components/TicketsPage/index.js
@@ -16,19 +16,17 @@ const TicketsPage = () => {
   const tickets = useSelector(state => Object.values(state.ticketsReducer))
   const events = useSelector(state => Object.values(state.eventsReducer))
 
-  const filtered = tickets.filter((ticket) => ticket.user_id === user.id)
+  const filtered = tickets.filter((ticket) => ticket?.user_id === user?.id)
   console.log(filtered)
-
-  if (user.id !== +userId) {
-    alert("Unauthorized [Incorrect User]");
-    history.push('/')
-    // window.location.reload();
-  }
 
   useEffect(() => {
     dispatch(allTickets(userId))
     dispatch(allEvents())
   }, [])
+
+  // if (user?.id !== +userId) {
+  //   return (<div> Incorrect User</div>)
+  // }
 
   const handleDelete = async (id) => {
     // openModal()
@@ -42,35 +40,46 @@ const TicketsPage = () => {
     // history.push(`/tickets/${user.id}`)
   }
   // const singleEvent = events?.find((event) => event.id === +eventId)
-  return (
-    <div className="ticket_container">
-      <h2>Your Tickets</h2>
-      {filtered?.map(ticket =>
-      (
-        <div className="ticket_div_container" key={ticket.id}>
-          <div className="ticket_div">
-            {events.filter((event) => event.id === ticket.event_id).map((singleEvent) => (
-              <Link to={`/events/${singleEvent?.id}`}>
-                <div className="single_event_container">
-                  <div className="single_event_div">
-                    {singleEvent?.pic_url.includes("jpeg") || singleEvent?.pic_url.includes("jpg") || singleEvent?.pic_url.includes("png") || singleEvent?.pic_url.includes("image") ? <img className="single_event_pic" src={singleEvent?.pic_url} /> : <img className="single_event_pic" src="https://static.thenounproject.com/png/340719-200.png" />}
-                    {/* <img className="single_event_pic" src={singleEvent?.pic_url} /> */}
+  console.log(user.id === +userId);
+  if (user?.id === +userId) {
+    return (
+      <div className="ticket_container">
+        <h2>Your Tickets</h2>
+        {filtered?.map(ticket =>
+        (
+          <div className="ticket_div_container" key={ticket.id}>
+            <div className="ticket_div">
+              {events.filter((event) => event.id === ticket.event_id).map((singleEvent) => (
+                <Link to={`/events/${singleEvent?.id}`}>
+                  <div className="single_event_container">
+                    <div className="single_event_div">
+                      {singleEvent?.pic_url.includes("jpeg") || singleEvent?.pic_url.includes("jpg") || singleEvent?.pic_url.includes("png") || singleEvent?.pic_url.includes("image") ? <img className="single_event_pic" src={singleEvent?.pic_url} /> : <img className="single_event_pic" src="https://static.thenounproject.com/png/340719-200.png" />}
+                      {/* <img className="single_event_pic" src={singleEvent?.pic_url} /> */}
+                    </div>
+                    <div className="single_event_text">
+                      <p>{singleEvent?.name}</p>
+                      <p>{singleEvent?.time.slice(0, -12)} <br /> {new Date(singleEvent?.time).getUTCHours() > 12 ? new Date(singleEvent?.time).getUTCHours() - 12 : new Date(singleEvent?.time).getUTCHours()}:{new Date(singleEvent?.time).getMinutes()}{new Date(singleEvent?.time).getUTCHours() > 12 ? "PM" : "AM"} </p>
+                      {/* <p>{singleEvent?.time}</p> */}
+                      <p>{singleEvent?.location}</p>
+                      <p className="single_event_text_num">Number of Tickets: {ticket.num_ticket}</p>
+                    </div>
                   </div>
-                  <div className="single_event_text">
-                    <p>{singleEvent?.name}</p>
-                    <p>{singleEvent?.time.slice(0, -12)} <br /> {new Date(singleEvent?.time).getUTCHours() > 12 ? new Date(singleEvent?.time).getUTCHours() - 12 : new Date(singleEvent?.time).getUTCHours()}:{new Date(singleEvent?.time).getMinutes()}{new Date(singleEvent?.time).getUTCHours() > 12 ? "PM" : "AM"} </p>
-                    {/* <p>{singleEvent?.time}</p> */}
-                    <p>{singleEvent?.location}</p>
-                    <p className="single_event_text_num">Number of Tickets: {ticket.num_ticket}</p>
-                  </div>
-                </div>
-              </Link>
-            ))}
-            <button className="single_event_button" onClick={() => handleDelete(ticket.id)}>Refund Ticket</button>
-          </div>
-        </div>))}
-    </div>
-  )
+                </Link>
+              ))}
+              <button className="single_event_button" onClick={() => handleDelete(ticket.id)}>Refund Ticket</button>
+            </div>
+          </div>))}
+      </div>
+    )
+  }
+  else {
+    return (
+      <div className="color_box">
+        <h2 >Your Tickets</h2>
+        <h3 className="wrong_user_h3">[UNAUTHORIZED] Incorrect User</h3>
+      </div>
+    )
+  }
 }
 
 export default TicketsPage
